style: naming for `ReturnYouTubeDislike` patch
