+++ src/main/kotlin/app/revanced/patches/youtube/layout/returnyoutubedislike/annotations/ReturnYouTubeDislikeCompatibility.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.youtube.layout.returnyoutubedislikes.annotations
+package app.revanced.patches.youtube.layout.returnyoutubedislike.annotations
 
 import app.revanced.patcher.annotation.Compatibility
 import app.revanced.patcher.annotation.Package
@@ -10,4 +10,4 @@ import app.revanced.patcher.annotation.Package
 )
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
-internal annotation class RYDCompatibility
+internal annotation class ReturnYouTubeDislikeCompatibility
