+++ src/main/kotlin/app/revanced/patches/youtube/layout/returnyoutubedislike/fingerprints/LikeFingerprint.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.youtube.layout.returnyoutubedislikes.fingerprints
+package app.revanced.patches.youtube.layout.returnyoutubedislike.fingerprints
 
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
@@ -6,7 +6,7 @@ import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.annotation.FuzzyPatternScanMethod
 import app.revanced.patcher.fingerprint.method.annotation.MatchingMethod
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-import app.revanced.patches.youtube.layout.returnyoutubedislikes.annotations.RYDCompatibility
+import app.revanced.patches.youtube.layout.returnyoutubedislike.annotations.ReturnYouTubeDislikeCompatibility
 import org.jf.dexlib2.AccessFlags
 
 @Name("like-fingerprint")
@@ -14,7 +14,7 @@ import org.jf.dexlib2.AccessFlags
     "Luqt;", "<init>"
 )
 @FuzzyPatternScanMethod(2)
-@RYDCompatibility
+@ReturnYouTubeDislikeCompatibility
 @Version("0.0.2")
 object LikeFingerprint : MethodFingerprint(
     "V",
