+++ src/main/kotlin/app/revanced/patches/youtube/layout/returnyoutubedislike/fingerprints/RemoveLikeFingerprint.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.youtube.layout.returnyoutubedislikes.fingerprints
+package app.revanced.patches.youtube.layout.returnyoutubedislike.fingerprints
 
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
@@ -6,20 +6,20 @@ import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.annotation.FuzzyPatternScanMethod
 import app.revanced.patcher.fingerprint.method.annotation.MatchingMethod
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-import app.revanced.patches.youtube.layout.returnyoutubedislikes.annotations.RYDCompatibility
+import app.revanced.patches.youtube.layout.returnyoutubedislike.annotations.ReturnYouTubeDislikeCompatibility
 import org.jf.dexlib2.AccessFlags
 
-@Name("dislike-fingerprint")
+@Name("remove-like-fingerprint")
 @MatchingMethod(
-    "Luqs;", "<init>"
+    "Luqw;", "<init>"
 )
 @FuzzyPatternScanMethod(2)
-@RYDCompatibility
+@ReturnYouTubeDislikeCompatibility
 @Version("0.0.2")
-object DislikeFingerprint : MethodFingerprint(
+object RemoveLikeFingerprint : MethodFingerprint(
     "V",
     AccessFlags.PROTECTED or AccessFlags.CONSTRUCTOR,
     null,
     null,
-    listOf("like/dislike")
+    listOf("like/removelike")
 )
