+++ src/main/kotlin/app/revanced/patches/youtube/layout/returnyoutubedislike/fingerprints/TextComponentSpecParentFingerprint.kt
@@ -1,18 +1,18 @@
-package app.revanced.patches.youtube.layout.returnyoutubedislikes.fingerprints
+package app.revanced.patches.youtube.layout.returnyoutubedislike.fingerprints
 
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
 import app.revanced.patcher.fingerprint.method.annotation.DirectPatternScanMethod
 import app.revanced.patcher.fingerprint.method.annotation.MatchingMethod
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-import app.revanced.patches.youtube.layout.returnyoutubedislikes.annotations.RYDCompatibility
+import app.revanced.patches.youtube.layout.returnyoutubedislike.annotations.ReturnYouTubeDislikeCompatibility
 
 @Name("text-component-spec-parent-fingerprint")
 @MatchingMethod(
     "Lnvy;", "e"
 )
 @DirectPatternScanMethod
-@RYDCompatibility
+@ReturnYouTubeDislikeCompatibility
 @Version("0.0.1")
 object TextComponentSpecParentFingerprint : MethodFingerprint(
     null,
